# Remove some extra string allocations in std.datetime.date.datetime