A micro optimization. No reason to allocate the string twice.