@JackStouffer 2017-05-23
Results

```
$ ldc2 -O -release test.d && ./test
old	5 secs, 902 ms, 131 μs, and 6 hnsecs
new	4 secs, 949 ms, 288 μs, and 6 hnsecs
```

Code

```d
import std.stdio;
import std.algorithm;
import std.conv;
import std.format;
import std.ascii;
import std.range;
import std.traits;
import std.string;
import std.datetime;
import std.utf;

enum testCount = 5_000_000;

string func1(S, R)(S date, R time)
{
    return format("%sT%s", date.toISOExtString(), time.toISOExtString());
}

string func2(S, R)(S date, R time)
{
    return format(
        "%sT%02d:%02d:%02d",
        date.toISOExtString(),
        time.hour,
        time.minute,
        time.second
    );
}

void main()
{
    immutable a = Date(2017, 1, 1);
    immutable b = TimeOfDay(12, 0, 0);

    auto result = to!Duration(benchmark!(() => func1(a, b))(testCount)[0]);
    auto result2 = to!Duration(benchmark!(() => func2(a, b))(testCount)[0]);

    writeln("old", "\t", result);
    writeln("new", "\t", result2);
}
```