@ZombineDev 2017-05-24
Thanks, try this one:
```D
string func3(S, R)(S date, R time)
{
    static void write02d(ref char[2] s, ubyte num)
    {
        assert (s == "00" && num < 100);
        if (num < 10)
            s[1] += num;
        else
        {
            s[0] += num / 10;
            s[1] += num % 10;
        }
    }

    void formatMDHMS(char[] start)
    {
        write02d(start[0 .. 2], date.month);
        start = start[3 .. $];
        write02d(start[0 .. 2], date.day);
        start = start[3 .. $];
        write02d(start[0 .. 2], time.hour);
        start = start[3 .. $];
        write02d(start[0 .. 2], time.minute);
        start = start[3 .. $];
        write02d(start[0 .. 2], time.second);
    }

    if (date.year < 0 || date.year >= 10_000)
    {
        char[15] result = "-00-00T00:00:00";
        formatMDHMS(result[1 .. $]);
        // too lazy to optimize this further :P
        return "%+d".format(date.year) ~ result;
    }
    else
    {
        char[19] result = "0000-00-00T00:00:00";
        auto y = date.year;
        auto p = &result[3];
        while (y)
        {
            *p-- += y % 10;
            y /= 10;
        }

        formatMDHMS(result[5 .. $]);
        return result.dup;
    }
}
```

:P