@wilzbach 2017-05-25
> Do you have any performance numbers?

@JackStouffer how about starting to collect all these benchmarks to a `/benchmark` folder [like druntime does](https://github.com/dlang/druntime/tree/master/benchmark), s.t. 
- people in the future have an easier time improving the performance
- we can start adding a CI for performance regressions