@JackStouffer 2017-05-25
Well, they're kind of different in scope. My benchmarks are about specific functions rather than measuring whole systems or storage types.