@JackStouffer 2017-06-28
@ZombineDev unfortunately, that's unsafe.