@ZombineDev 2017-06-29
@JackStouffer while it's not `@safe`, I don't think it is unsafe. Care to explain in more detail? Is there some case that I have missed in my solution?