@JackStouffer 2017-06-29
The problem is the current function is `@safe`, so making it `@system` would be breaking.