@JackStouffer 2017-07-05
@jmdavis Fixed

Also changed to the new compile time format strings.